+++ doc/changelog.rst
@@ -9,3 +9,5 @@ v2.0.0
 * breaking feature: Variables like ``$this`` or ``${THIS}`` are substituted in
   all scalar in yaml file, if you relied on this kind of values present in the
   config verbatim, pass ``vars=None`` to config parser
+* feature: Add ``--print-config-vars`` command-line argument to print variables
+  used in specific config